requirements.txt
openai
huggingface_hub
gradio
python-dotenv 